Add read_data script
The script reads the data of a split (train/val/test)
from the TFRecords.

It processes
the data as ((image, caption), slid_caption) tuples:

the first inner tuple (image, caption) contains the image
and the full caption associated to
the image, that has to be provided
as input of the model.

The slid_caption is the same caption
slid of one poistion to the left, corresponding
to the output we expect the model to
provide when using teacher forcing. The
elements of the slid caption are also
decreased of -1 so that the output
neuron 0 will have sense (and not
be referred to the 0 padding) and
so we can ignore the -1s in the loss
function (which correspond to
padding tokens).